requirements.txt
antlr4-python2-runtime==4.5.3
blinker==1.4
cffi==1.9.1
click==6.6
cryptography==1.7
elasticsearch==5.0.1
enum34==1.1.6
Flask==0.11.1
frozendict==0.6
gevent==1.1.2
gevent-websocket==0.9.5
greenlet==0.4.11
idna==2.1
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
Markdown==2.6.7
MarkupSafe==0.23
mongoengine==0.11.0
passlib==1.7.0
pyasn1==0.1.9
pycparser==2.17
pymongo==3.4.0
pyOpenSSL==16.2.0
python-dateutil==2.5.0
python-engineio==0.9.0
python-socketio==1.2
PyYAML==3.12
requests==2.12.4
six==1.10.0
sortedcontainers==1.5.3
splunk-sdk==1.6.1
urllib3==1.19.1
Werkzeug==0.11.11
